# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proposal Engine Public API",
    "version": "1.0.0",
    "description": "REST API for integrating Proposal Engine with external systems. Supports audit creation, proposal retrieval, outreach triggering, client management, and webhook callbacks.",
    "contact": {
      "email": "api@proposalengine.io"
    }
  },
  "servers": [
    { "url": "/api/v1", "description": "Production" }
  ],
  "security": [{ "BearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key prefixed with `pe_pub_`. Obtain from your agency dashboard."
      }
    },
...
  },
  "paths": {
...
    "/webhooks/{id}": {
      "delete": {
        "summary": "Remove a webhook endpoint",
        "operationId": "deleteWebhook",
        "tags": ["Webhooks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}